Speed Demo (iPhone and Apple Watch App)
This app demonstrates the CoreLocation and Watch Connectivity features of the Apple Watch.  The watch app tracks the current speed of the watch and sends that data over to the iPhone counterpart app.  An alert on the phone goes off when the watch reaches a certain speed, which can be configured by the user on the phone app.